Add initial networking code.
radiod already knows what radios it has spawned a rigctld process for,
and it also already knows about all the potential radios on the system.
Thus, it makes sense that radiod should have a network port where
clients could connect and ask about what radios are present and perhaps
get information as they come and go.

This commit adds the skeleton for that - listening on ports, sending and
receiving messages, etc.  It does not yet do anything useful, however.

I am also using async and safe-exceptions now to make the code nicer.
In particular, I got rid of that dumb loop function and the associated
mvar.